feat(kuma-cp) removed dataplane token server
