+++ app/kumactl/pkg/tokens/client_test.go
@@ -3,8 +3,9 @@ package tokens_test
 import (
 	"fmt"
 	"github.com/Kong/kuma/app/kumactl/pkg/tokens"
+	admin_server "github.com/Kong/kuma/pkg/admin-server"
+	admin_server_config "github.com/Kong/kuma/pkg/config/admin-server"
 	config_kumactl "github.com/Kong/kuma/pkg/config/app/kumactl/v1alpha1"
-	token_server "github.com/Kong/kuma/pkg/config/token-server"
 	"github.com/Kong/kuma/pkg/core/xds"
 	"github.com/Kong/kuma/pkg/sds/auth"
 	"github.com/Kong/kuma/pkg/test"
@@ -45,23 +46,23 @@ var _ = Describe("Tokens Client", func() {
 		Expect(err).ToNot(HaveOccurred())
 		publicPort = p
 
-		srv := server.DataplaneTokenServer{
-			Config: &token_server.DataplaneTokenServerConfig{
+		adminCfg := admin_server_config.AdminServerConfig{
+			DataplaneTokenWs: &admin_server_config.DataplaneTokenWsConfig{
 				Enabled: true,
-				Local: &token_server.LocalDataplaneTokenServerConfig{
-					Port: uint32(port),
-				},
-				Public: &token_server.PublicDataplaneTokenServerConfig{
-					Enabled:        true,
-					Port:           uint32(publicPort),
-					Interface:      "localhost",
-					TlsCertFile:    filepath.Join("..", "..", "..", "..", "pkg", "tokens", "builtin", "server", "testdata", "server-cert.pem"),
-					TlsKeyFile:     filepath.Join("..", "..", "..", "..", "pkg", "tokens", "builtin", "server", "testdata", "server-key.pem"),
-					ClientCertsDir: filepath.Join("..", "..", "..", "..", "pkg", "tokens", "builtin", "server", "testdata", "authorized-clients"),
-				},
 			},
-			Issuer: &staticTokenIssuer{},
+			Local: &admin_server_config.LocalAdminServerConfig{
+				Port: uint32(port),
+			},
+			Public: &admin_server_config.PublicAdminServerConfig{
+				Enabled:        true,
+				Port:           uint32(publicPort),
+				Interface:      "localhost",
+				TlsCertFile:    filepath.Join("..", "..", "..", "..", "pkg", "admin-server", "testdata", "server-cert.pem"),
+				TlsKeyFile:     filepath.Join("..", "..", "..", "..", "pkg", "admin-server", "testdata", "server-key.pem"),
+				ClientCertsDir: filepath.Join("..", "..", "..", "..", "pkg", "admin-server", "testdata", "authorized-clients"),
+			},
 		}
+		srv := admin_server.NewAdminServer(adminCfg, server.NewWebservice(&staticTokenIssuer{}))
 
 		ch := make(chan struct{})
 		errCh := make(chan error)
@@ -105,8 +106,8 @@ var _ = Describe("Tokens Client", func() {
 				return fmt.Sprintf("https://localhost:%d", publicPort)
 			},
 			config: &config_kumactl.Context_DataplaneTokenApiCredentials{
-				ClientCert: filepath.Join("..", "..", "..", "..", "pkg", "tokens", "builtin", "server", "testdata", "authorized-client-cert.pem"),
-				ClientKey:  filepath.Join("..", "..", "..", "..", "pkg", "tokens", "builtin", "server", "testdata", "authorized-client-key.pem"),
+				ClientCert: filepath.Join("..", "..", "..", "..", "pkg", "admin-server", "testdata", "authorized-client-cert.pem"),
+				ClientKey:  filepath.Join("..", "..", "..", "..", "pkg", "admin-server", "testdata", "authorized-client-key.pem"),
 			},
 		}),
 	)
