+++ pkg/admin-server/server.go
@@ -2,6 +2,8 @@ package admin_server
 
 import (
 	"context"
+	"crypto/tls"
+	"crypto/x509"
 	"fmt"
 	admin_server "github.com/Kong/kuma/pkg/config/admin-server"
 	config_core "github.com/Kong/kuma/pkg/config/core"
@@ -13,7 +15,10 @@ import (
 	"github.com/emicklei/go-restful"
 	"github.com/pkg/errors"
 	"go.uber.org/multierr"
+	"io/ioutil"
 	"net/http"
+	"path/filepath"
+	"strings"
 )
 
 var (
@@ -39,6 +44,14 @@ func NewAdminServer(cfg admin_server.AdminServerConfig, services ...*restful.Web
 func (a *AdminServer) Start(stop <-chan struct{}) error {
 	httpServer, httpErrChan := a.startHttpServer()
 
+	var httpsServer *http.Server
+	var httpsErrChan chan error
+	if a.cfg.Public.Enabled {
+		httpsServer, httpsErrChan = a.startHttpsServer()
+	} else {
+		httpsErrChan = make(chan error)
+	}
+
 	select {
 	case <-stop:
 		log.Info("stopping")
@@ -46,8 +59,15 @@ func (a *AdminServer) Start(stop <-chan struct{}) error {
 		if err := httpServer.Shutdown(context.Background()); err != nil {
 			multiErr = multierr.Combine(err)
 		}
+		if httpsServer != nil {
+			if err := httpsServer.Shutdown(context.Background()); err != nil {
+				multiErr = multierr.Combine(err)
+			}
+		}
 		return multiErr
 	case err := <-httpErrChan:
+		return err
+	case err := <-httpsErrChan:
 		return err
 	}
 }
@@ -75,6 +95,63 @@ func (a *AdminServer) startHttpServer() (*http.Server, chan error) {
 	return server, errChan
 }
 
+func (a *AdminServer) startHttpsServer() (*http.Server, chan error) {
+	errChan := make(chan error)
+
+	tlsConfig, err := requireClientCerts(a.cfg.Public.ClientCertsDir)
+	if err != nil {
+		errChan <- err
+	}
+
+	server := &http.Server{
+		Addr:      fmt.Sprintf("%s:%d", a.cfg.Public.Interface, a.cfg.Public.Port),
+		Handler:   a.container,
+		TLSConfig: tlsConfig,
+	}
+
+	go func() {
+		defer close(errChan)
+		if err := server.ListenAndServeTLS(a.cfg.Public.TlsCertFile, a.cfg.Public.TlsKeyFile); err != nil {
+			if err != http.ErrServerClosed {
+				log.Error(err, "https server terminated with an error")
+				errChan <- err
+				return
+			}
+		}
+		log.Info("https server terminated normally")
+	}()
+	log.Info("starting server", "interface", a.cfg.Public.Interface, "port", a.cfg.Public.Port, "tls", true)
+	return server, errChan
+}
+
+func requireClientCerts(certsDir string) (*tls.Config, error) {
+	files, err := ioutil.ReadDir(certsDir)
+	if err != nil {
+		return nil, err
+	}
+	clientCertPool := x509.NewCertPool()
+	for _, file := range files {
+		if file.IsDir() {
+			continue
+		}
+		if !strings.HasSuffix(file.Name(), ".pem") {
+			continue
+		}
+		path := filepath.Join(certsDir, file.Name())
+		caCert, err := ioutil.ReadFile(path)
+		if err != nil {
+			return nil, errors.Wrapf(err, "could not read certificate %s", path)
+		}
+		clientCertPool.AppendCertsFromPEM(caCert)
+	}
+	tlsConfig := &tls.Config{
+		ClientCAs:  clientCertPool,
+		ClientAuth: tls.RequireAndVerifyClientCert,
+	}
+	tlsConfig.BuildNameToCertificate()
+	return tlsConfig, nil
+}
+
 func SetupServer(rt runtime.Runtime) error {
 	var webservices []*restful.WebService
 
@@ -95,7 +172,7 @@ func SetupServer(rt runtime.Runtime) error {
 
 func dataplaneTokenWs(rt runtime.Runtime) (*restful.WebService, error) {
 	if !rt.Config().AdminServer.DataplaneTokenWs.Enabled {
-		log.Info("Dataplane Token Webservice disabled")
+		log.Info("Dataplane Token Webservice is disabled. Dataplane Tokens won't be verified.")
 		return nil, nil
 	}
