+++ pkg/admin-server/server_suite_test.go
@@ -1,0 +1,13 @@
+package admin_server_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestAdminServer(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Admin Server Suite")
+}
