+++ pkg/admin-server/testdata/authorized-client-cert.pem
