+++ pkg/admin-server/testdata/authorized-clients/authorized-client-cert.pem
