+++ pkg/admin-server/testdata/authorized-clients/file-that-should-be-ignored
